BGDIINF_SB-2160: first steps on menu merge Moving all menu related component to a dedicated menu/ component folder. Fix some z-index positioning Fix some (collapse) animation not triggering (because behind a v-if instead of a v-show)
