+++ src/App.vue
@@ -1,6 +1,7 @@
 <template>
-    <div id="app">
+    <div id="main-component">
         <router-view />
+        <div id="toolbox"></div>
     </div>
 </template>
 
@@ -43,11 +44,21 @@ export default {
 // this import needs to happen only once, otherwise bootstrap is import/added
 // to the output CSS as many time as this file is imported
 @import 'node_modules/bootstrap/scss/bootstrap';
-#app {
+#main-component {
     font-family: Avenir, Helvetica, Arial, sans-serif;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
     text-align: center;
     color: #2c3e50;
+}
+#toolbox {
+    position: absolute;
+    z-index: $zindex-toolbox;
+    width: 100%;
+    height: 100%;
+    pointer-events: none;
+    & > * {
+        pointer-events: all;
+    }
 }
 </style>
