+++ src/modules/map/components/MapFooter.vue
@@ -2,7 +2,7 @@
     <div id="footer" class="align-items-center">
         <div id="footer-target-scale"></div>
         <div id="footer-target-mouse" class="flex-grow-1"></div>
-        <div v-if="showAppVersion" class="mx-3 app-version small">v{{ appVersion }}</div>
+        <div v-if="showAppVersion" class="mx-3 app-version small pe-none">v{{ appVersion }}</div>
         <a
             :href="`https://www.geo.admin.ch/${$i18n.locale}/about-swiss-geoportal/impressum.html#copyright`"
             target="_blank"
