+++ src/modules/menu/components/menu/MenuSettings.vue
