+++ src/modules/toolbox/ToolboxModule.vue
@@ -1,19 +1,19 @@
 <template>
-    <div id="toolbox">
+    <teleport to="#toolbox">
         <transition name="slide-right">
             <div
                 v-if="showToolbox"
                 class="right-toolbox"
                 :class="{ 'currently-drawing': isCurrentlyDrawing }"
             >
+                <GeolocButton class="geoloc-button mb-1" />
                 <ZoomButtons class="toolbox-zoom-buttons" />
-                <GeolocButton class="geoloc-button" />
             </div>
         </transition>
         <transition name="slide-left">
             <BackgroundSelectorButton v-if="showToolbox" class="toolbox-bg-buttons" />
         </transition>
-    </div>
+    </teleport>
 </template>
 
 <script>
@@ -37,36 +37,15 @@ export default {
 @import 'src/scss/media-query.mixin';
 @import 'src/scss/variables';
 
-#toolbox {
-    .right-toolbox,
-    .toolbox-bg-buttons {
-        position: absolute;
-        z-index: $zindex-footer;
-    }
-    .right-toolbox {
-        top: $header-height + $screen-padding-for-ui-elements;
-        right: $screen-padding-for-ui-elements;
-        &.currently-drawing {
-            top: calc(124px + #{$screen-padding-for-ui-elements});
-        }
-        .toolbox-zoom-buttons {
-            position: relative;
-        }
-    }
-    .toolbox-bg-buttons {
-        left: $screen-padding-for-ui-elements;
-        bottom: $footer-height + $screen-padding-for-ui-elements;
-    }
+.right-toolbox {
+    float: right;
+    position: relative;
+    margin: $screen-padding-for-ui-elements;
 }
-@include respond-above(lg) {
-    #toolbox {
-        .right-toolbox {
-            top: 2 * $header-height + $screen-padding-for-ui-elements;
-            &.currently-drawing {
-                top: calc(124px + #{$screen-padding-for-ui-elements});
-            }
-        }
-    }
+.toolbox-bg-buttons {
+    position: absolute;
+    left: $screen-padding-for-ui-elements;
+    bottom: $footer-height + $screen-padding-for-ui-elements;
 }
 // transition definition
 .slide-left-leave-active,
